Ensure /debug/state does not access mid-tick data
